pyproject.toml
[project]
name = "trading-bot-ai"
version = "2.0.0"
description = "A SOLID-principles based cryptocurrency trading bot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "binance-connector==3.6.0",
    "pandas==2.2.2",
    "pandas-ta==0.3.14b",
    "setuptools==69.5.1",
    "numpy==1.26.4",
    "python-dotenv==1.0.0",
    "python-binance>=1.0.29",
]

[project.scripts]
trading-bot = "main:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "freezegun>=1.2.0",
    "responses>=0.23.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
test = [
    "freezegun>=1.5.5",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "responses>=0.25.8",
]
